requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==4.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.36
flask-socketio==5.3.6
simple-websocket==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
